[Update #1] separate main() from main.swift to GitGrassIndicator.swift
